requirements.txt
llama-index-core>=0.12.52
llama-index-llms-openai>=0.4.7
llama-index-llms-anthropic>=0.7.6
llama-index-llms-ollama>=0.6.2
llama-index-llms-gemini>=0.5.0
llama-index-llms-mistralai>=0.6.1
llama-index-llms-groq>=0.3.2
llama-index-vector-stores-chroma>=0.4.1
llama-index-embeddings-openai>=0.3.1
llama-index-embeddings-huggingface>=0.5.5
llama-index>=0.12.52
chromadb
streamlit>=1.24.0
beautifulsoup4
youtube-transcript-api
python-dotenv
pandas
requests
importlib-metadata
watchdog
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
